# backend/app/schemas/schemas.py
"""Pydantic schemas for request/response validation."""

from pydantic import BaseModel, EmailStr, Field, ConfigDict
from typing import Optional, List
from datetime import datetime


# ─────────────────────────────────────────
# Auth Schemas
# ─────────────────────────────────────────

class UserRegister(BaseModel):
    email: EmailStr
    full_name: str = Field(..., min_length=2, max_length=255)
    password: str = Field(..., min_length=8, max_length=128)


class UserLogin(BaseModel):
    email: EmailStr
    password: str


class UserResponse(BaseModel):
    model_config = ConfigDict(from_attributes=True)

    id: str
    email: str
    full_name: str
    role: str
    is_active: bool
    created_at: datetime


class TokenResponse(BaseModel):
    access_token: str
    token_type: str = "bearer"
    user: UserResponse


# ─────────────────────────────────────────
# Document Schemas
# ─────────────────────────────────────────

class ExtractedFieldSchema(BaseModel):
    model_config = ConfigDict(from_attributes=True)

    id: str
    field_name: str
    field_value: Optional[str] = None
    field_category: Optional[str] = None


class SummarySchema(BaseModel):
    model_config = ConfigDict(from_attributes=True)

    id: str
    summary_text: str
    coverage_summary: Optional[str] = None
    exclusions_summary: Optional[str] = None
    waiting_period_summary: Optional[str] = None
    premium_summary: Optional[str] = None
    model_used: str
    created_at: datetime


class RiskAnalysisSchema(BaseModel):
    model_config = ConfigDict(from_attributes=True)

    id: str
    clause_text: str
    risk_type: str
    severity: str
    explanation: Optional[str] = None
    recommendation: Optional[str] = None
    created_at: datetime


class DocumentResponse(BaseModel):
    model_config = ConfigDict(from_attributes=True)

    id: str
    original_filename: str
    file_type: str
    file_size_bytes: int
    page_count: int
    status: str
    extraction_method: Optional[str] = None
    created_at: datetime
    updated_at: datetime


class DocumentDetailResponse(DocumentResponse):
    extracted_text: Optional[str] = None
    summary: Optional[SummarySchema] = None
    extracted_fields: List[ExtractedFieldSchema] = []
    risk_analyses: List[RiskAnalysisSchema] = []


# ─────────────────────────────────────────
# AI Service Schemas
# ─────────────────────────────────────────

class SummarizeRequest(BaseModel):
    document_id: str


class ExtractFieldsRequest(BaseModel):
    document_id: str


class RiskAnalysisRequest(BaseModel):
    document_id: str


class ExtractedFieldsResponse(BaseModel):
    document_id: str
    fields: List[ExtractedFieldSchema]


class RiskAnalysisResponse(BaseModel):
    document_id: str
    risks: List[RiskAnalysisSchema]
    overall_risk_level: str


class SummarizeResponse(BaseModel):
    document_id: str
    summary: SummarySchema


# ─────────────────────────────────────────
# Comparison Schemas
# ─────────────────────────────────────────

class CompareRequest(BaseModel):
    document_ids: List[str]


class FeatureWinnerSchema(BaseModel):
    feature: str
    winner: str
    reason: str


class ComparisonSynthesisSchema(BaseModel):
    synthesis: str
    best_for: str
    verdict: str
    feature_winners: List[FeatureWinnerSchema] = []


class CompareResponse(BaseModel):
    documents: List[DocumentDetailResponse]
    comparison_synthesis: ComparisonSynthesisSchema