# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"Healthcare Insurance Document Intelligence System","description":"AI-powered platform for analyzing healthcare insurance documents","version":"1.0.0"},"paths":{"/api/v1/auth/register":{"post":{"tags":["Authentication"],"summary":"Register","description":"Register a new user account.","operationId":"register_api_v1_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegister"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["Authentication"],"summary":"Login","description":"Login with email and password.","operationId":"login_api_v1_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/me":{"get":{"tags":["Authentication"],"summary":"Get Me","description":"Get current user profile.","operationId":"get_me_api_v1_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/documents/upload":{"post":{"tags":["Documents"],"summary":"Upload Document","description":"Upload a healthcare insurance document (PDF or image).","operationId":"upload_document_api_v1_documents_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_document_api_v1_documents_upload_post"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocumentResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/documents":{"get":{"tags":["Documents"],"summary":"List Documents","description":"List all documents for the current user.","operationId":"list_documents_api_v1_documents_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/DocumentResponse"},"type":"array","title":"Response List Documents Api V1 Documents Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/documents/{document_id}":{"get":{"tags":["Documents"],"summary":"Get Document","description":"Get detailed document info including AI results.","operationId":"get_document_api_v1_documents__document_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"document_id","in":"path","required":true,"schema":{"type":"string","title":"Document Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocumentDetailResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Documents"],"summary":"Delete Document","description":"Delete a document and its associated data.","operationId":"delete_document_api_v1_documents__document_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"document_id","in":"path","required":true,"schema":{"type":"string","title":"Document Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/documents/compare":{"post":{"tags":["Documents"],"summary":"Compare Documents","description":"Compare multiple health insurance policies side-by-side.","operationId":"compare_documents_api_v1_documents_compare_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompareRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompareResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/ai/summarize":{"post":{"tags":["AI Services"],"summary":"Summarize Document","description":"Generate an AI summary of the insurance document.","operationId":"summarize_document_api_v1_ai_summarize_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SummarizeRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SummarizeResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/ai/extract-fields":{"post":{"tags":["AI Services"],"summary":"Extract Fields","description":"Extract key policy fields from the insurance document.","operationId":"extract_fields_api_v1_ai_extract_fields_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExtractFieldsRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExtractedFieldsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/ai/risk-analysis":{"post":{"tags":["AI Services"],"summary":"Risk Analysis","description":"Detect and analyze risky clauses in the insurance document.","operationId":"risk_analysis_api_v1_ai_risk_analysis_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RiskAnalysisRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RiskAnalysisResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/":{"get":{"tags":["Health"],"summary":"Root","description":"Health check endpoint.","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/health":{"get":{"tags":["Health"],"summary":"Health Check","description":"Detailed health check.","operationId":"health_check_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_upload_document_api_v1_documents_upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_document_api_v1_documents_upload_post"},"CompareRequest":{"properties":{"document_ids":{"items":{"type":"string"},"type":"array","maxItems":5,"minItems":1,"title":"Document Ids"}},"type":"object","required":["document_ids"],"title":"CompareRequest"},"CompareResponse":{"properties":{"documents":{"items":{"$ref":"#/components/schemas/DocumentDetailResponse"},"type":"array","title":"Documents"},"comparison":{"$ref":"#/components/schemas/ComparisonSynthesis"}},"type":"object","required":["documents","comparison"],"title":"CompareResponse"},"ComparisonSynthesis":{"properties":{"synthesis":{"type":"string","title":"Synthesis"},"best_for":{"type":"string","title":"Best For"},"verdict":{"type":"string","title":"Verdict"}},"type":"object","required":["synthesis","best_for","verdict"],"title":"ComparisonSynthesis"},"DocumentDetailResponse":{"properties":{"id":{"type":"string","title":"Id"},"original_filename":{"type":"string","title":"Original Filename"},"file_type":{"type":"string","title":"File Type"},"file_size_bytes":{"type":"integer","title":"File Size Bytes"},"page_count":{"type":"integer","title":"Page Count"},"status":{"type":"string","title":"Status"},"extraction_method":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Extraction Method"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"},"extracted_text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Extracted Text"},"summary":{"anyOf":[{"$ref":"#/components/schemas/SummarySchema"},{"type":"null"}]},"extracted_fields":{"items":{"$ref":"#/components/schemas/ExtractedFieldSchema"},"type":"array","title":"Extracted Fields","default":[]},"risk_analyses":{"items":{"$ref":"#/components/schemas/RiskAnalysisSchema"},"type":"array","title":"Risk Analyses","default":[]}},"type":"object","required":["id","original_filename","file_type","file_size_bytes","page_count","status","created_at","updated_at"],"title":"DocumentDetailResponse"},"DocumentResponse":{"properties":{"id":{"type":"string","title":"Id"},"original_filename":{"type":"string","title":"Original Filename"},"file_type":{"type":"string","title":"File Type"},"file_size_bytes":{"type":"integer","title":"File Size Bytes"},"page_count":{"type":"integer","title":"Page Count"},"status":{"type":"string","title":"Status"},"extraction_method":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Extraction Method"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","original_filename","file_type","file_size_bytes","page_count","status","created_at","updated_at"],"title":"DocumentResponse"},"ExtractFieldsRequest":{"properties":{"document_id":{"type":"string","title":"Document Id"}},"type":"object","required":["document_id"],"title":"ExtractFieldsRequest"},"ExtractedFieldSchema":{"properties":{"id":{"type":"string","title":"Id"},"field_name":{"type":"string","title":"Field Name"},"field_value":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Field Value"},"field_category":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Field Category"}},"type":"object","required":["id","field_name"],"title":"ExtractedFieldSchema"},"ExtractedFieldsResponse":{"properties":{"document_id":{"type":"string","title":"Document Id"},"fields":{"items":{"$ref":"#/components/schemas/ExtractedFieldSchema"},"type":"array","title":"Fields"}},"type":"object","required":["document_id","fields"],"title":"ExtractedFieldsResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"RiskAnalysisRequest":{"properties":{"document_id":{"type":"string","title":"Document Id"}},"type":"object","required":["document_id"],"title":"RiskAnalysisRequest"},"RiskAnalysisResponse":{"properties":{"document_id":{"type":"string","title":"Document Id"},"risks":{"items":{"$ref":"#/components/schemas/RiskAnalysisSchema"},"type":"array","title":"Risks"},"overall_risk_level":{"type":"string","title":"Overall Risk Level"}},"type":"object","required":["document_id","risks","overall_risk_level"],"title":"RiskAnalysisResponse"},"RiskAnalysisSchema":{"properties":{"id":{"type":"string","title":"Id"},"clause_text":{"type":"string","title":"Clause Text"},"risk_type":{"type":"string","title":"Risk Type"},"severity":{"type":"string","title":"Severity"},"explanation":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Explanation"},"recommendation":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Recommendation"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","clause_text","risk_type","severity","created_at"],"title":"RiskAnalysisSchema"},"SummarizeRequest":{"properties":{"document_id":{"type":"string","title":"Document Id"}},"type":"object","required":["document_id"],"title":"SummarizeRequest"},"SummarizeResponse":{"properties":{"document_id":{"type":"string","title":"Document Id"},"summary":{"$ref":"#/components/schemas/SummarySchema"}},"type":"object","required":["document_id","summary"],"title":"SummarizeResponse"},"SummarySchema":{"properties":{"id":{"type":"string","title":"Id"},"summary_text":{"type":"string","title":"Summary Text"},"coverage_summary":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Coverage Summary"},"exclusions_summary":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Exclusions Summary"},"waiting_period_summary":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Waiting Period Summary"},"premium_summary":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Premium Summary"},"model_used":{"type":"string","title":"Model Used"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","summary_text","model_used","created_at"],"title":"SummarySchema"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"},"user":{"$ref":"#/components/schemas/UserResponse"}},"type":"object","required":["access_token","user"],"title":"TokenResponse"},"UserLogin":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserLogin"},"UserRegister":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"full_name":{"type":"string","maxLength":255,"minLength":2,"title":"Full Name"},"password":{"type":"string","maxLength":128,"minLength":8,"title":"Password"}},"type":"object","required":["email","full_name","password"],"title":"UserRegister"},"UserResponse":{"properties":{"id":{"type":"string","title":"Id"},"email":{"type":"string","title":"Email"},"full_name":{"type":"string","title":"Full Name"},"role":{"type":"string","title":"Role"},"is_active":{"type":"boolean","title":"Is Active"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","email","full_name","role","is_active","created_at"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/v1/auth/login"}}}}}}